requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
SQLAlchemy==2.0.43
Pillow==10.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
Flask-APScheduler==1.13.1
email-validator==2.3.0
gunicorn==22.0.0
